Replaced REAL*8 with REAL.

+++ Declarations.f90
@@ -289,7 +289,7 @@
 
 ! Note that the variable names in this COMMON block are not consistent.
 
-      DOUBLE PRECISION XOPT,       C,       FOPT
+      REAL XOPT,       C,       FOPT
       COMMON /sacmn /  XOPT(MVAR), C(MVAR), FOPT
 
 ! MVAR = 100 (the variable formerly also known as NMAX, almost subtly different from MPAR)
